pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "music-charts"
version = "1.0.0"
description = "Data pipeline for analyzing track performance across streaming platforms"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12, <3.15"
authors = [
    { name = "Dylan \"Dyl-M\" Monfret", email = "dyl_m.dev@proton.me" }
]
keywords = ["music", "analytics", "spotify", "youtube", "streaming"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

dependencies = [
    "google-auth>=2.41.0, <2.42.0",
    "google-auth-oauthlib>=1.2.3",
    "libpybee>=0.2.1",
    "odfpy>=1.4.1",
    "pandas>=2.3.0",
    "plotly>=6.5.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-youtube>=0.9.7",
    "requests>=2.32.5",
    "scikit-learn>=1.7.0",
    "tenacity>=9.0.0",
    "tqdm>=4.67.0",
    "typer>=0.15.0",
]

[project.optional-dependencies]
dev = [
    "pandas-stubs>=2.3.3",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-requests>=2.32.0",
]

[project.scripts]
msc = "msc.cli:app"

[project.urls]
Homepage = "https://github.com/Dyl-M/music-charts"
Repository = "https://github.com/Dyl-M/music-charts"
Issues = "https://github.com/Dyl-M/music-charts/issues"

[tool.hatch.build.targets.wheel]
packages = ["msc"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["msc"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "libpybee.*",
    "plotly.*",
    "pyyoutube.*",
    "sklearn.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["_tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["msc"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "pragma: nocover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@(abc\\.)?abstractmethod",
    "def __repr__",
    "if __name__ == .__main__.:",
]
